BugFixed - Unit moves again after defeating the oponent
